moved initial_setup to main
